windows compile update

+++ src/Entry.cpp
@@ -24,7 +24,7 @@ void listDevices() {
 
     LOGP_I("Listing Interfaces .....");
 
-    for (int j = 0; j < deviceList->getCount(); j++) {
+    for (size_t j = 0; j < deviceList->getCount(); j++) {
         LOGP_I("%s : %s", CommInterfaceType::getName(deviceList->get(j)->getType()),
                deviceList->get(j)->getName().c_str());
     }
